requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
opencv-python==4.0.0.21
Pillow==6.0.0
pyparsing==2.3.1
python-dateutil==2.8.0
six==1.12.0
unittest-xml-reporting==2.5.1
